pyproject.toml
[project]
name = "slidegen"
version = "0.1.0"
description = "PPT Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "pydantic[email]>=2.8.0",
    "python-dotenv>=1.0.1",
    "SQLAlchemy>=2.0.31",
    "aiomysql>=0.2.0",
    "PyMySQL>=1.1.1",
    "celery>=5.4.0",
    "redis>=5.0.7",
    "python-multipart>=0.0.9",
    "Jinja2>=3.1.4",
    "httpx>=0.27.0",
    "loguru>=0.7.2",
    "openpyxl>=3.1.5",
    "xlrd>=2.0.1",
    "puremagic",
    "python-pptx",
    "types-pytz",
    "mammoth",
    "bs4",
    "markdownify==0.14.1",
    "pandas",
    "pydantic-settings>=2.8.1",
    "rich>=14.0.0",
    "gmssl>=3.2.2",
    "pytest>=8.3.5",
    "inflect>=7.5.0",
    "mypy>=1.15.0",
    "openai>=1.72.0",
    "mdurl>=0.1.2",
    "more-itertools>=10.6.0",
    "tenacity>=9.1.2",
    "pandas-stubs>=2.2.3.250527",
    "python-jose[cryptography]>=3.5.0",
    "sse-starlette>=2.3.6",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.26.0",
    "watchfiles>=0.22.0",
    "pytest>=7.0.0",
    "ruff>=0.11.4",
    "pre-commit>=4.2.0",
]


[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic", "tests"]
ignore_missing_imports = true
disallow_untyped_defs = true
warn_return_any = false

[tool.ruff]
target-version = "py312"
exclude = ["alembic", "scripts", "tests"]
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # import
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "C416", # use list comprehension instead of map
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

# 使用中科大源和清华源
# [[tool.uv.index]]
# name = "ustc"
# url = "https://mirrors.ustc.edu.cn/pypi/simple"

# [[tool.uv.index]]
# name = "tsinghua"
# url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"

[tool.uv]
package = false
